Backport OS X fix from audacious player.
Dynamically linked loadable modules are have the ".so" suffix on Darwin,
too. This is not to be confused with dynamically linked loadable
*libraries*, which have the ".dylib" suffix on Mac OS X/Darwin.
As Audacious uses dyn. MODULES, we should use the ".so" extension (which
is also defined by glib-2.0 for dyn. modules.)

Also, employ the rpath and install_name fixes from audacious player.