New configuration window
A configuration window has been implemented. Its function is to check if authenticated requests can be made to last.fm. If so, the plugin is configured. If not, the user must authorize Audacious on his last.fm account and then re-check the application on the plugin.
Some code was cleaned up here and there and removed some debug code.