Fixes "Audacious needs five seconds to quit with enabled Scrobbler2 plugin", reported by radioactiveman.
This was happening because the communicating thread slept in intervals of 7 seconds, and re-checked
at that interval to see if scrobbling was enabled. Scrobbling gets enabled as soon as a profile is
configured and a working Internet connection is available. So, if one of these wasn't true, the
user would see a delay when shutting down the plugin.

Fixed by using pthread_cond_timedwait as per jlindgren's suggestion. Thanks.