scrobbler2: Started working on migration from the old scrobbler.
This adds the ability to upgrade from the old plugin without having to reconfigure the Scrobbler.
A best-effort attempt is made to retrieve a session key. If this is not possible, the plugin will
spit a warning message so that the user knows why is Audacious not scrobbling anymore.

Not done yet:
 - The Scrobbler2 plugin should be automagically enabled if the scrobbler plugin was enabled (now this is trivial);
 - The tracks left on the scrobblerqueue.txt are not being scrobbled yet.