Add support for re-scrobbling old tracks.
This also paves the way for #260.
See:
 * http://redmine.audacious-media-player.org/issues/343#note-24
 * http://redmine.audacious-media-player.org/issues/260#note-1

Tracks ignored by last.fm because they are too old (currently, 2 weeks) will be
resubmitted with the timestamp of when the ignored status was returned.

As we're running against time with last.fm (see issue comments linked above),
this is a quick and dirty implementation. The code can be improved (mostly
refactors and function names), but for now this will do.

Manually tested with scrobbler.log contents:
 - 1 old track
 - 1 old track + 1 recent track
 - 1 old track + many recent tracks
 - many old tracks
 - many old tracks + 1 recent track
 - many old tracks + many recent tracks
 - 1 recent track
 - many recent tracks