qtui: Let Qt on Linux determine the toolbar icon size.
With several themes available on Linux, hard-coding the icon size can
cause blurry icons. One affected icon theme is for example Faenza.

When not setting the size manually, Qt determines the size by the
application's style. [1] This fixes the appearance because 24px icons
are used directly and not downscaled to 22px.

[1] http://doc.qt.io/qt-5/qtoolbar.html#iconSize-prop