Only extend the icon theme search path if necessary
Check if our QtUi "edit-find" icon is available (when no icon theme is
set, i.e. chiefly on stock Mac or MS Windows). It will be if the QtUi
theme is included in the embedded Qt resource. If not, add the likely
location where the theme lives on disk to the path.

The fallback isn't required and can be removed after the embedded theme
has been tested sufficiently (and/or the theme is no longer installed
in AudPath::DataDir).