statusicon-qt: Don't show/hide the UI on Mac.
Apparently opening the statusicon menu on Mac also causes a Trigger
event.  This is inconsistent with Qt behavior on X11 (where Trigger
and Context are distinct), so it seems like a Qt bug, but we'll deal
with it.

Note, Q_OS_MACOS is not available yet in Qt 5.2 so until we require
a higher Qt version, we'll stick with Q_OS_MAC.  Q_OS_OSX could be
used if we really needed to distinguish between OS X/macOS and iOS,
but we don't support iOS anyway, so meh.