Add a libopenmpt-based tracker module plugin.
This plugin should support all files that are suppored by the
modplug-based plugin. However, due to how widespread libmodplug is, its
plugin is not removed, but instead set at a slightly lower priority, so
that if both it and the libopenmpt-based plugin are installed,
libopenmpt will be preferred. This allows systems without libopenmpt to
still have module playback support.