Add Disc Number support
Although ID3 and a number of tagging schemes do not support disc numbers - Disc numbers are can be an integral part of tagging for multi-disc albums, especially Classical music.

This PR adds Disc number support to audacious-plugins - there is a
complementary PR to audacious adding support there.

This implements Feature #603

Signed-off-by: Andrew Thornton <art27@cantab.net>