lyricwiki-qt: Add support for chartlyrics.com. Closes: #1217
lyrics.ovh has been unreliable in the past months.
Meanwhile it is working again but still has issues at times.

Offer chartlyrics.com as an alternative. It is also not perfect
with no support for HTTPS and lyrics for new songs are missing.
Hoever having a choice is still better. And it is free for
non-commercial use, unlike many other sites.

See also:
- http://www.chartlyrics.com/api.aspx (Terms of Use)
- https://github.com/NTag/lyrics.ovh/issues/15
- https://github.com/NTag/lyrics.ovh/issues/17