Make UI appear on Qt::ApplicationActive for MacOS
If the status icon plugin is used with close to system tray
the UI main window is closed. On Mac OS there is not an easy
way to make it reappear. The system tray activation trigger
is disabled on Mac OS. This is for good reason due to one
button mouse support and that this usually is intended to
open the status/tray menu. Instead we can handle the
application state change activate signal which will fire
when the dock icon is clicked.